+++ py-rustitude/src/dataset.rs
@@ -28,7 +28,7 @@ impl Event {
         self.0.index
     }
     #[getter]
-    fn weight(&self) -> f64 {
+    fn weight(&self) -> f32 {
         self.0.weight
     }
     #[getter]
@@ -49,7 +49,7 @@ impl Event {
             .collect()
     }
     #[getter]
-    fn eps(&self) -> [f64; 3] {
+    fn eps(&self) -> [f32; 3] {
         [self.0.eps[0], self.0.eps[1], self.0.eps[2]]
     }
     fn __str__(&self) -> String {
@@ -87,7 +87,7 @@ impl Dataset {
         self.0.events.iter().cloned().map(Event::from).collect()
     }
     #[getter]
-    fn weights(&self) -> Vec<f64> {
+    fn weights(&self) -> Vec<f32> {
         self.0.weights()
     }
     fn __len__(&self) -> PyResult<usize> {
@@ -101,7 +101,7 @@ impl Dataset {
     #[pyo3(signature = (range, bins, daughter_indices=None))]
     fn split_m(
         &self,
-        range: (f64, f64),
+        range: (f32, f32),
         bins: usize,
         daughter_indices: Option<Vec<usize>>,
     ) -> (Vec<Vec<usize>>, Vec<usize>, Vec<usize>) {
@@ -119,27 +119,27 @@ impl Dataset {
 
     #[staticmethod]
     fn from_dict(py: Python, data: HashMap<String, PyObject>) -> PyResult<Self> {
-        let e_beam_vec: Vec<f64> = data["E_Beam"].extract(py)?;
-        let px_beam_vec: Vec<f64> = data["Px_Beam"].extract(py)?;
-        let py_beam_vec: Vec<f64> = data["Py_Beam"].extract(py)?;
-        let pz_beam_vec: Vec<f64> = data["Pz_Beam"].extract(py)?;
-        let weight_vec: Vec<f64> = data
+        let e_beam_vec: Vec<f32> = data["E_Beam"].extract(py)?;
+        let px_beam_vec: Vec<f32> = data["Px_Beam"].extract(py)?;
+        let py_beam_vec: Vec<f32> = data["Py_Beam"].extract(py)?;
+        let pz_beam_vec: Vec<f32> = data["Pz_Beam"].extract(py)?;
+        let weight_vec: Vec<f32> = data
             .get("Weight")
             .map_or_else(|| Ok(vec![1.0; e_beam_vec.len()]), |obj| obj.extract(py))?;
-        let eps_vec: Vec<Vector3<f64>> = data.get("EPS").map_or_else(
+        let eps_vec: Vec<Vector3<f32>> = data.get("EPS").map_or_else(
             || Ok(vec![Vector3::default(); e_beam_vec.len()]),
             |obj| {
-                obj.extract::<Vec<Vec<f64>>>(py).map(|vvf: Vec<Vec<f64>>| {
+                obj.extract::<Vec<Vec<f32>>>(py).map(|vvf: Vec<Vec<f32>>| {
                     vvf.into_iter()
                         .map(Vector3::from_vec)
-                        .collect::<Vec<Vector3<f64>>>()
+                        .collect::<Vec<Vector3<f32>>>()
                 })
             },
         )?;
-        let e_finalstate_vec: Vec<Vec<f64>> = data["E_FinalState"].extract(py)?;
-        let px_finalstate_vec: Vec<Vec<f64>> = data["Px_FinalState"].extract(py)?;
-        let py_finalstate_vec: Vec<Vec<f64>> = data["Py_FinalState"].extract(py)?;
-        let pz_finalstate_vec: Vec<Vec<f64>> = data["Pz_FinalState"].extract(py)?;
+        let e_finalstate_vec: Vec<Vec<f32>> = data["E_FinalState"].extract(py)?;
+        let px_finalstate_vec: Vec<Vec<f32>> = data["Px_FinalState"].extract(py)?;
+        let py_finalstate_vec: Vec<Vec<f32>> = data["Py_FinalState"].extract(py)?;
+        let pz_finalstate_vec: Vec<Vec<f32>> = data["Pz_FinalState"].extract(py)?;
         Ok(Self(rust::Dataset::new(
             (
                 e_beam_vec,
@@ -211,7 +211,7 @@ impl Dataset {
             .map_err(PyErr::from)
     }
     #[staticmethod]
-    fn from_parquet_with_eps(path: &str, eps: Vec<f64>) -> PyResult<Self> {
+    fn from_parquet_with_eps(path: &str, eps: Vec<f32>) -> PyResult<Self> {
         rust::Dataset::from_parquet_with_eps(path, eps)
             .map(Dataset::from)
             .map_err(PyErr::from)
